make async requests separately than creating view model
