+++ src/Orckestra.Composer.Search/Orckestra.Composer.Search.csproj
@@ -82,9 +82,11 @@
     <Compile Include="Facets\SelectedFacet.cs" />
     <Compile Include="Factory\FacetFactory.cs" />
     <Compile Include="Factory\FacetPredicateFactory.cs" />
+    <Compile Include="Factory\IProductSearchViewModelFactory.cs" />
     <Compile Include="Factory\ISelectedFacetFactory.cs" />
     <Compile Include="Factory\IFacetFactory.cs" />
     <Compile Include="Factory\IFacetPredicateFactory.cs" />
+    <Compile Include="Factory\ProductSearchViewModelFactory.cs" />
     <Compile Include="Factory\SelectedFacetFactory.cs" />
     <Compile Include="Providers\IAutocompleteProvider.cs" />
     <Compile Include="Providers\ISearchTermsTransformationProvider.cs" />
