+++ src/Orckestra.Composer.Search/Services/BaseSearchViewService.cs
@@ -3,8 +3,6 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using System.Threading.Tasks;
-using Orckestra.Composer.Configuration;
-using Orckestra.Composer.Enums;
 using Orckestra.Composer.Parameters;
 using Orckestra.Composer.Providers;
 using Orckestra.Composer.Providers.Dam;
@@ -19,8 +17,6 @@ using Orckestra.Composer.Search.ViewModels;
 using Orckestra.Composer.Services;
 using Orckestra.Composer.Utils;
 using Orckestra.Composer.ViewModels;
-using Orckestra.Overture.ServiceModel;
-using Orckestra.Overture.ServiceModel.Products.Inventory;
 using Orckestra.Overture.ServiceModel.Search;
 using Facet = Orckestra.Composer.Search.Facets.Facet;
 using SearchFilter = Orckestra.Composer.Parameters.SearchFilter;
@@ -40,45 +36,36 @@ namespace Orckestra.Composer.Search.Services
         protected IDamProvider DamProvider { get; }
         protected IFacetFactory FacetFactory { get; }
         protected ILocalizationProvider LocalizationProvider { get; }
-        protected IProductUrlProvider ProductUrlProvider { get; }
         protected ISearchRepository SearchRepository { get; }
         protected ISearchUrlProvider SearchUrlProvider { get; }
         protected ISelectedFacetFactory SelectedFacetFactory { get; }
-        protected IViewModelMapper ViewModelMapper { get; }
         protected IPriceProvider PriceProvider { get; }
         protected IComposerContext ComposerContext { get; }
         protected IProductSettingsViewService ProductSettings { get; }
-        protected IScopeViewService ScopeViewService { get; }
-        protected IRecurringOrdersSettings RecurringOrdersSettings { get; private set; }
+        protected IProductSearchViewModelFactory ProductSearchViewModelFactory { get; private set;  }
 
         protected BaseSearchViewService(
             ISearchRepository searchRepository,
-            IViewModelMapper viewModelMapper,
             IDamProvider damProvider,
             ILocalizationProvider localizationProvider,
-            IProductUrlProvider productUrlProvider,
             ISearchUrlProvider searchUrlProvider,
             IFacetFactory facetFactory,
             ISelectedFacetFactory selectedFacetFactory,
             IPriceProvider priceProvider,
             IComposerContext composerContext,
             IProductSettingsViewService productSettings,
-            IScopeViewService scopeViewService,
-            IRecurringOrdersSettings recurringOrdersSettings)
+            IProductSearchViewModelFactory productSearchViewModelFactory)
         {
             SearchRepository = searchRepository ?? throw new ArgumentNullException(nameof(searchRepository));
-            ViewModelMapper = viewModelMapper ?? throw new ArgumentNullException(nameof(viewModelMapper));
             DamProvider = damProvider ?? throw new ArgumentNullException(nameof(damProvider));
             LocalizationProvider = localizationProvider ?? throw new ArgumentNullException(nameof(localizationProvider));
-            ProductUrlProvider = productUrlProvider ?? throw new ArgumentNullException(nameof(productUrlProvider));
             SearchUrlProvider = searchUrlProvider ?? throw new ArgumentNullException(nameof(searchUrlProvider));
             SelectedFacetFactory = selectedFacetFactory ?? throw new ArgumentNullException(nameof(selectedFacetFactory));
             FacetFactory = facetFactory ?? throw new ArgumentNullException(nameof(facetFactory));
             PriceProvider = priceProvider ?? throw new ArgumentNullException(nameof(priceProvider));
             ComposerContext = composerContext ?? throw new ArgumentNullException(nameof(composerContext));
             ProductSettings = productSettings ?? throw new ArgumentNullException(nameof(productSettings));
-            ScopeViewService = scopeViewService ?? throw new ArgumentNullException(nameof(scopeViewService));
-            RecurringOrdersSettings = recurringOrdersSettings ?? throw new ArgumentNullException(nameof(recurringOrdersSettings));
+            ProductSearchViewModelFactory = productSearchViewModelFactory ?? throw new ArgumentNullException(nameof(productSearchViewModelFactory));
         }
 
         protected virtual IList<Facet> BuildFacets(SearchCriteria criteria, ProductSearchResult searchResult)
@@ -186,14 +173,15 @@ namespace Orckestra.Composer.Search.Services
             }
 
             var imgDictionary = LineItemHelper.BuildImageDictionaryFor(param.ImageUrls);
-
+            var searchResultsList = new List<(ProductSearchViewModel, ProductDocument)>();
 
             // Populate search results
             foreach (var resultItem in param.SearchResult.Documents)
             {
-                var productSearchVm = await CreateProductSearchViewModelAsync(resultItem, param, imgDictionary).ConfigureAwait(false);
-                searchResultViewModel.SearchResults.Add(productSearchVm);
+                searchResultsList.Add((ProductSearchViewModelFactory.GetProductSearchViewModel(resultItem, param.SearchParam.Criteria, imgDictionary), resultItem));
             }
+            
+            searchResultViewModel.SearchResults = await AppendProductSearchViewModels(searchResultsList);
 
             var facets = BuildFacets(param.SearchParam.Criteria, param.SearchResult);
             searchResultViewModel.Facets = facets;
@@ -214,65 +202,22 @@ namespace Orckestra.Composer.Search.Services
             return SearchConfiguration.SearchSortBy.Where(d => !d.SearchType.HasValue || d.SearchType.Value == searchType).ToList();
         }
 
-        protected virtual async Task<ProductSearchViewModel> CreateProductSearchViewModelAsync(ProductDocument productDocument, CreateProductSearchResultsViewModelParam<TParam> createSearchViewModelParam, IDictionary<Tuple<string, string>, ProductMainImage> imgDictionary)
+
+        // NOTE: when fetching data for products from OCC APIs, make sure to query data in batches for optimal performance
+        // https://docs.orckestra.com/developer-documentation/platform-performance/batch-api-requests
+        protected virtual async Task<IList<ProductSearchViewModel>> AppendProductSearchViewModels(IList<(ProductSearchViewModel, ProductDocument)> productSearchResultList)
         {
-            var cultureInfo = createSearchViewModelParam.SearchParam.Criteria.CultureInfo;
-
-            string variantId = null;
-            if (productDocument.PropertyBag.ContainsKey(VariantPropertyBagKey))
-            {
-                variantId = productDocument.PropertyBag[VariantPropertyBagKey] as string;
-            }
-
-            var productSearchVm = ViewModelMapper.MapTo<ProductSearchViewModel>(productDocument, cultureInfo);
-            productSearchVm.ProductId = productDocument.ProductId;
-            productSearchVm.BrandId = ExtractLookupId("Brand_Facet", productDocument.PropertyBag);
-            MapProductSearchViewModelInfos(productSearchVm, productDocument, cultureInfo);
-            MapProductSearchViewModelUrl(productSearchVm, variantId, cultureInfo, createSearchViewModelParam.SearchParam.Criteria.BaseUrl);
-            MapProductSearchViewModelImage(productSearchVm, imgDictionary);
-            productSearchVm.IsAvailableToSell = await GetProductSearchViewModelAvailableForSell(productSearchVm, productDocument).ConfigureAwait(false);
-            var pricing = await PriceProvider.GetPriceAsync(productSearchVm.HasVariants, productDocument).ConfigureAwait(false);
-            MapProductSearchViewModelPricing(productSearchVm, pricing);
-            productSearchVm.IsRecurringOrderEligible = RecurringOrdersSettings.Enabled && productDocument.PropertyBag.IsRecurringOrderEligible();
-
-            productSearchVm.Context["IsRecurringOrderEligible "] = productSearchVm.IsRecurringOrderEligible;
-
-            return productSearchVm;
-        }
-
-        protected virtual string ExtractLookupId(string fieldName, PropertyBag propertyBag)
-        {
-            if (propertyBag == null) { return null; }
-            var fieldValue = propertyBag.ContainsKey(fieldName) 
-                ? (propertyBag[fieldName] as string ?? (propertyBag[fieldName] as string[])[0])
-                : null;
-            if (string.IsNullOrWhiteSpace(fieldValue)) { return null; }
-
-            var extractedValues = fieldValue.Split(new[] { "::" }, StringSplitOptions.None);
-            return extractedValues.Length < 3
-                ? null
-                : extractedValues[2];
-        }
-
-        protected virtual async Task<bool> GetProductSearchViewModelAvailableForSell(
-            ProductSearchViewModel productSearchViewModel,
-            ProductDocument productDocument)
-        {
-            if (productSearchViewModel.HasVariants) { return true; }
-
             _productSettings = await ProductSettings.GetProductSettings(ComposerContext.Scope, ComposerContext.CultureInfo).ConfigureAwait(false);
 
-            if (!_productSettings.IsInventoryEnabled) { return true; }
+            foreach (var (productSearchVm, productDocument) in productSearchResultList)
+            {
+                if (!_productSettings.IsInventoryEnabled) { productSearchVm.IsAvailableToSell = true; }
 
-            var availableStatusesForSell = ComposerConfiguration.AvailableStatusForSell;
+                var pricing = await PriceProvider.GetPriceAsync(productSearchVm.HasVariants, productDocument).ConfigureAwait(false);
+                MapProductSearchViewModelPricing(productSearchVm, pricing);
+            }
 
-            return (from inventoryItemAvailability
-                        in productDocument.InventoryLocationStatuses
-                    from inventoryItemStatuse
-                        in inventoryItemAvailability.Statuses
-                    select GetInventoryItemStatus(inventoryItemStatuse.Status))
-                    .Any(inventoryItemStatus => availableStatusesForSell
-                        .Any(availableStatusForSell => availableStatusForSell == inventoryItemStatus));
+            return productSearchResultList.Select((resultItem) => resultItem.Item1).ToList();
         }
 
         protected virtual SelectedFacets FlattenFilterList(IList<SearchFilter> filters, CultureInfo cultureInfo)
@@ -292,23 +237,6 @@ namespace Orckestra.Composer.Search.Services
                 Facets = facets,
                 IsAllRemovable = facets.Count(filter => filter.IsRemovable) > 1
             };
-        }
-
-        protected static InventoryStatusEnum GetInventoryItemStatus(InventoryStatus inventoryStatus)
-        {
-            switch (inventoryStatus)
-            {
-                case InventoryStatus.InStock:
-                    return InventoryStatusEnum.InStock;
-                case InventoryStatus.OutOfStock:
-                    return InventoryStatusEnum.OutOfStock;
-                case InventoryStatus.PreOrder:
-                    return InventoryStatusEnum.PreOrder;
-                case InventoryStatus.BackOrder:
-                    return InventoryStatusEnum.BackOrder;
-                default:
-                    return InventoryStatusEnum.Unspecified;
-            }
         }
 
         protected abstract string GenerateUrl(CreateSearchPaginationParam<TParam> param);
@@ -430,25 +358,6 @@ namespace Orckestra.Composer.Search.Services
             return previousPage;
         }
 
-        protected static bool HasVariants(ProductDocument resultItem)
-        {
-            if (resultItem?.PropertyBag == null)
-            {
-                return false;
-            }
-
-            if (!resultItem.PropertyBag.TryGetValue("GroupCount", out object variantCountObject) || variantCountObject == null)
-            {
-                return false;
-            }
-
-            var variantCountString = variantCountObject.ToString();
-
-            int.TryParse(variantCountString, out int result);
-
-            return result > 0;
-        }
-
         protected virtual void MapProductSearchViewModelPricing(ProductSearchViewModel productSearchVm, ProductPriceSearchViewModel pricing)
         {
             productSearchVm.DisplayListPrice = pricing.DisplayPrice;
@@ -459,60 +368,7 @@ namespace Orckestra.Composer.Search.Services
             productSearchVm.IsOnSale = pricing.IsOnSale;
             productSearchVm.PriceListId = pricing.PriceListId;
         }
-        protected virtual void MapProductSearchViewModelImage(
-            ProductSearchViewModel productSearchVm,
-            IDictionary<Tuple<string, string>,
-            ProductMainImage> imgDictionary)
-        {
-            string productVariantId = null;
-            if (!string.IsNullOrWhiteSpace(productSearchVm.VariantId))
-            {
-                productVariantId = productSearchVm.VariantId;
-            }
-
-            var imageKey = Tuple.Create(productSearchVm.ProductId, productVariantId);
-            var imageExists = imgDictionary.TryGetValue(imageKey, out ProductMainImage mainImage);
-
-            if (imageExists)
-            {
-                productSearchVm.ImageUrl = mainImage.ImageUrl;
-                productSearchVm.FallbackImageUrl = mainImage.FallbackImageUrl;
-            }
-        }
-
-        protected virtual void MapProductSearchViewModelInfos(
-            ProductSearchViewModel productSearchVm,
-            ProductDocument productDocument,
-            CultureInfo cultureInfo)
-        {
-            productSearchVm.DisplayName = TrimProductDisplayName(productSearchVm.FullDisplayName);
-            productSearchVm.Description = null; // We don't need Description in Search Results, setting to null, to reduce HTML size
-
-
-            //TODO use ProductDocument property when overture will have add it.
-            if (productSearchVm.Bag.ContainsKey("DefinitionName"))
-            {
-                productSearchVm.DefinitionName = productSearchVm.Bag["DefinitionName"].ToString();
-            }
-
-            productSearchVm.HasVariants = HasVariants(productDocument);
-        }
-
-        protected virtual void MapProductSearchViewModelUrl(
-            ProductSearchViewModel productSearchVm,
-            string productVariantId,
-            CultureInfo cultureInfo, string baseUrl)
-        {
-            productSearchVm.Url = ProductUrlProvider.GetProductUrl(new GetProductUrlParam
-            {
-                CultureInfo = cultureInfo,
-                ProductId = productSearchVm.ProductId,
-                ProductName = productSearchVm.FullDisplayName,
-                VariantId = productVariantId,
-                BaseUrl = baseUrl,
-                SKU = productSearchVm.Sku
-            });
-        }
+      
 
         /// <summary>
         ///     Searches the available products based on the given search criteria.
@@ -562,16 +418,6 @@ namespace Orckestra.Composer.Search.Services
                         PropertyBag = document.PropertyBag
                     }).ToList()
             };
-        }
-
-        private static string TrimProductDisplayName(string displayName)
-        {
-            if (string.IsNullOrWhiteSpace(displayName)) { return string.Empty; }
-
-            var trimmedDisplayName = displayName.Substring(
-                0, Math.Min(displayName.Length, DisplayConfiguration.ProductNameMaxLength));
-
-            return trimmedDisplayName;
         }
     }
 }
